requirements.txt
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-openai
llama-parse
qdrant-client
python-dotenv
pydantic
pydantic-settings
pytest
pandas
requests
datasets
sec-edgar-downloader
fastembed
tqdm
